+++ Client/app/mercadopago/components/MercadoPagoButton/index.tsx
@@ -20,7 +20,7 @@ export const MercadoPagoButton = ({ product }: MercadoPagoButtonProps) => {
       setLoading(true);
 
       try {
-        const { data: preference } = await axios.post("../../api/checkout.ts",
+        const { data: preference } = await axios.post("/api/checkout.ts",
           product
         );
