+++ Client/app/page.tsx
@@ -6,6 +6,8 @@ import Footer from "@/components/footer/Footer"
 export default function Home() {
   return (
     <div>
+
+      <script src="https://sdk.mercadopago.com/js/v2"></script>
       <HomeSection />
       <ServiceSection />
