--- Server/.env.example
@@ -1,4 +1,0 @@
-PORT = 
-MONGODB_URL = 
-EMAIL = 
-EMAIL_PASSWORD =
