+++ Server/package.json
@@ -27,11 +27,13 @@
     "express": "^4.18.2",
     "express-session": "^1.17.3",
     "jsonwebtoken": "^9.0.0",
+    "mercadopago": "^1.5.16",
     "mock-aws-s3": "^4.0.2",
     "mongodb": "^5.6.0",
     "mongoose": "^7.2.4",
     "morgan": "^1.10.0",
     "next": "^13.4.7",
+    "next-auth": "^4.22.1",
     "passport": "^0.6.0",
     "passport-local": "^1.0.0",
     "postcss": "^8.4.24",
@@ -40,7 +42,6 @@
     "tailwindcss": "^3.3.2",
     "typescript": "^5.1.3",
     "vercel": "^31.0.1",
-    "next-auth": "^4.22.1",
     "zod-error": "^1.5.0"
   },
   "devDependencies": {
